+++ tests/testthat/test-as-mcmc.R
@@ -14,8 +14,8 @@ test_that("as_mcmc.nlist", {
     as_mcmc(nlist(x = 1)),
     structure(
       1,
-      .Dim = c(1L, 1L),
-      .Dimnames = list(NULL, "x"),
+      dim = c(1L, 1L),
+      dimnames = list(NULL, "x"),
       mcpar = c(
         1,
         1,
@@ -28,8 +28,8 @@ test_that("as_mcmc.nlist", {
     as_mcmc(nlist(x = matrix(1:6, 2))),
     structure(
       1:6,
-      .Dim = c(1L, 6L),
-      .Dimnames = list(
+      dim = c(1L, 6L),
+      dimnames = list(
         NULL,
         c(
           "x[1,1]",
@@ -52,8 +52,8 @@ test_that("as_mcmc.nlist", {
     as_mcmc(nlist(x = 1:2, y = 3)),
     structure(
       c(1, 2, 3),
-      .Dim = c(1L, 3L),
-      .Dimnames = list(
+      dim = c(1L, 3L),
+      dimnames = list(
         NULL,
         c("x[1]", "x[2]", "y")
       ),
@@ -65,8 +65,8 @@ test_that("as_mcmc.nlist", {
     as_mcmc(nlist(x = NA_real_, y = 1)),
     structure(
       c(NA, 1),
-      .Dim = 1:2,
-      .Dimnames = list(
+      dim = 1:2,
+      dimnames = list(
         NULL,
         c(
           "x",
@@ -82,8 +82,8 @@ test_that("as_mcmc.nlist", {
     as_mcmc(nlist::nlist(x = 1, y = matrix(1:4, 2))),
     structure(
       c(1, 1, 2, 3, 4),
-      .Dim = c(1L, 5L),
-      .Dimnames = list(
+      dim = c(1L, 5L),
+      dimnames = list(
         NULL,
         c("x", "y[1,1]", "y[2,1]", "y[1,2]", "y[2,2]")
       ),
@@ -106,8 +106,8 @@ test_that("as_mcmc.nlists", {
     as_mcmc(nlists(nlist())),
     structure(
       numeric(0),
-      .Dim = 1:0,
-      .Dimnames = list(NULL, NULL),
+      dim = 1:0,
+      dimnames = list(NULL, NULL),
       mcpar = c(
         1,
         1,
@@ -120,8 +120,8 @@ test_that("as_mcmc.nlists", {
     as_mcmc(nlists(nlist(), nlist())),
     structure(
       numeric(0),
-      .Dim = c(2L, 0L),
-      .Dimnames = list(
+      dim = c(2L, 0L),
+      dimnames = list(
         NULL,
         NULL
       ),
@@ -134,8 +134,8 @@ test_that("as_mcmc.nlists", {
     as_mcmc(nlists(nlist(x = 1))),
     structure(
       1,
-      .Dim = c(1L, 1L),
-      .Dimnames = list(NULL, "x"),
+      dim = c(1L, 1L),
+      dimnames = list(NULL, "x"),
       mcpar = c(
         1,
         1,
@@ -148,8 +148,8 @@ test_that("as_mcmc.nlists", {
     as_mcmc(nlists(nlist(x = 1), nlist(x = 3))),
     structure(
       c(1, 3),
-      .Dim = 2:1,
-      .Dimnames = list(NULL, "x"),
+      dim = 2:1,
+      dimnames = list(NULL, "x"),
       mcpar = c(
         1,
         2,
@@ -162,8 +162,8 @@ test_that("as_mcmc.nlists", {
     as_mcmc(nlists(nlist(x = matrix(1:6, 2)))),
     structure(
       1:6,
-      .Dim = c(1L, 6L),
-      .Dimnames = list(
+      dim = c(1L, 6L),
+      dimnames = list(
         NULL,
         c(
           "x[1,1]",
@@ -190,11 +190,11 @@ test_that("as_mcmc.nlists", {
     )),
     structure(
       c(1L, 3L, 2L, 4L, 3L, 5L, 4L, 6L, 5L, 7L, 6L, 8L),
-      .Dim = c(
+      dim = c(
         2L,
         6L
       ),
-      .Dimnames = list(
+      dimnames = list(
         NULL,
         c(
           "x[1,1]",
@@ -220,8 +220,8 @@ test_that("as.mcmc.nlist", {
     as.mcmc(nlist(x = 1)),
     structure(
       1,
-      .Dim = c(1L, 1L),
-      .Dimnames = list(NULL, "x"),
+      dim = c(1L, 1L),
+      dimnames = list(NULL, "x"),
       mcpar = c(
         1,
         1,
@@ -234,8 +234,8 @@ test_that("as.mcmc.nlist", {
     as.mcmc(nlist(x = matrix(1:6, 2))),
     structure(
       1:6,
-      .Dim = c(1L, 6L),
-      .Dimnames = list(
+      dim = c(1L, 6L),
+      dimnames = list(
         NULL,
         c(
           "x[1,1]",
@@ -258,8 +258,8 @@ test_that("as.mcmc.nlist", {
     as.mcmc(nlist(x = 1:2, y = 3)),
     structure(
       c(1, 2, 3),
-      .Dim = c(1L, 3L),
-      .Dimnames = list(
+      dim = c(1L, 3L),
+      dimnames = list(
         NULL,
         c("x[1]", "x[2]", "y")
       ),
@@ -271,8 +271,8 @@ test_that("as.mcmc.nlist", {
     as.mcmc(nlist(x = NA_real_, y = 1)),
     structure(
       c(NA, 1),
-      .Dim = 1:2,
-      .Dimnames = list(
+      dim = 1:2,
+      dimnames = list(
         NULL,
         c(
           "x",
@@ -288,8 +288,8 @@ test_that("as.mcmc.nlist", {
     as.mcmc(nlist::nlist(x = 1, y = matrix(1:4, 2))),
     structure(
       c(1, 1, 2, 3, 4),
-      .Dim = c(1L, 5L),
-      .Dimnames = list(
+      dim = c(1L, 5L),
+      dimnames = list(
         NULL,
         c("x", "y[1,1]", "y[2,1]", "y[1,2]", "y[2,2]")
       ),
@@ -312,8 +312,8 @@ test_that("as.mcmc.nlists", {
     as.mcmc(nlists(nlist())),
     structure(
       numeric(0),
-      .Dim = 1:0,
-      .Dimnames = list(NULL, NULL),
+      dim = 1:0,
+      dimnames = list(NULL, NULL),
       mcpar = c(
         1,
         1,
@@ -326,8 +326,8 @@ test_that("as.mcmc.nlists", {
     as.mcmc(nlists(nlist(), nlist())),
     structure(
       numeric(0),
-      .Dim = c(2L, 0L),
-      .Dimnames = list(
+      dim = c(2L, 0L),
+      dimnames = list(
         NULL,
         NULL
       ),
@@ -340,8 +340,8 @@ test_that("as.mcmc.nlists", {
     as.mcmc(nlists(nlist(x = 1))),
     structure(
       1,
-      .Dim = c(1L, 1L),
-      .Dimnames = list(NULL, "x"),
+      dim = c(1L, 1L),
+      dimnames = list(NULL, "x"),
       mcpar = c(
         1,
         1,
@@ -354,8 +354,8 @@ test_that("as.mcmc.nlists", {
     as.mcmc(nlists(nlist(x = 1), nlist(x = 3))),
     structure(
       c(1, 3),
-      .Dim = 2:1,
-      .Dimnames = list(NULL, "x"),
+      dim = 2:1,
+      dimnames = list(NULL, "x"),
       mcpar = c(
         1,
         2,
@@ -368,8 +368,8 @@ test_that("as.mcmc.nlists", {
     as.mcmc(nlists(nlist(x = matrix(1:6, 2)))),
     structure(
       1:6,
-      .Dim = c(1L, 6L),
-      .Dimnames = list(
+      dim = c(1L, 6L),
+      dimnames = list(
         NULL,
         c(
           "x[1,1]",
@@ -396,11 +396,11 @@ test_that("as.mcmc.nlists", {
     )),
     structure(
       c(1L, 3L, 2L, 4L, 3L, 5L, 4L, 6L, 5L, 7L, 6L, 8L),
-      .Dim = c(
+      dim = c(
         2L,
         6L
       ),
-      .Dimnames = list(
+      dimnames = list(
         NULL,
         c(
           "x[1,1]",